Split payload operations
- Refactored `payload_operations.py` by moving `Node`, `AlgorithmNode`, `ProbeNode`, and `Pipeline` to dedicated modules
    - Introduced `nodes.py` and `pipeline.py` to improve modularity of payload operations
- Fixed typos in `context_types.py` and `image_data_types.py`, correcting `Collecton` to `Collection`
